+++ python-wrapper/src/neo4j_viz/node.py
@@ -98,3 +98,10 @@ class Node(
         by_field = [v.validation_alias.choices for k, v in Node.model_fields.items() if k not in exempted_fields]  # type: ignore
 
         return {str(alias) for aliases in by_field for alias in aliases}
+
+    @staticmethod
+    def basic_fields_validation_aliases() -> set[str]:
+        mandatory_fields = ["id", "caption"]
+        by_field = [v.validation_alias.choices for k, v in Node.model_fields.items() if k in mandatory_fields]  # type: ignore
+
+        return {str(alias) for aliases in by_field for alias in aliases}