+++ python-wrapper/src/neo4j_viz/relationship.py
@@ -109,3 +109,14 @@ class Relationship(
         ]
 
         return {str(alias) for aliases in by_field for alias in aliases}
+
+    @staticmethod
+    def basic_fields_validation_aliases() -> set[str]:
+        basic_fields = ["id", "source", "target", "caption"]
+        by_field = [
+            v.validation_alias.choices  # type: ignore
+            for k, v in Relationship.model_fields.items()
+            if k in basic_fields
+        ]
+
+        return {str(alias) for aliases in by_field for alias in aliases}