+++ python-wrapper/src/neo4j_viz/visualization_graph.py
@@ -22,6 +22,7 @@ from .options import (
 from .relationship import Relationship
 
 
+# TODO helper for map properties to fields. helper for set caption (simplicity), expand doc string
 class VisualizationGraph:
     """
     A graph to visualize.